Upload files from SmugMug to Nixplay
Now files that are not present in the Nixplay album will be downloaded
from SmugMug and uploaded to Nixplay.  Still need to add them to the
Nixplay "playlist", then save the playlist (sigh).  Also not deleting
photos from Nixplay that are no longer in the SmugMug album (though
that's less important)