Add --max parameter so that we can sync more than 100 pictures
Problem: The SmugMug API defaults to returning the first 100 pictures,
so after the album has more than 100 pictures in it, the syncing stops
adding new photos.  (And it would probably delete some already existing
photos if deleting was implemented)

Solution: SmugMug API supports larger return sets (?count=1000) and also
pagination (?start=1001).  This commit doesn't implement pagination
because it's not clear that's a big win (we'd still have to load the
entirety of the smugmug album and nixplay album into working set in
order to do the set difference anyway).  But, set the count to a
user-specificed "--max" parameter.