Refactor smugmug-specific login activities to subcommand and pkg
Problem: A few smugmug-specific things were in high-level picsync
pkgs and cmds.  Since it was the only complicated login flow before,
this was a little unclean but livable.  Adding google photo support
make the need for separation more clear.

Solution: Refactor all the login stuff into the smugmug package.  Make
the smugmug auth structures more sensible because we're going to be
working with them alongside the google photo creds too.