Add Google Photos OAuth2 Workflow
Problem: No way to auth to Google Photos.

Solution: Add "picsync googlephotos login" command, which uses the
OAuth2 workflow to sign in to the Google Photos API.  This uses the
"Desktop Application" version of the workflow.

The app key/token have to be provided via environment variables/config
(PICSYNC_GOOGLEPHOTOS_API_KEY and PICSYNC_GOOGLEPHOTOS_API_SECRET).
These are not actually super-secret but this is an unapproved temporary
app so I'm not committing them anywhere yet.