Add beginning of SQLite3 driver
Problem: Google Photos does not provide an MD5 checksum for photos you've stored.  Nixplay only provides an MD5.  So there's no direct way to tell "is this photo in googlephotos already in nixplay?".

Solution: Create a cache of the metadata we need.  For now we'll use md5s
since we can get that direct from Nixplay API.  The cache is a sqlite database, driven by the modernc.org/sqlite project, which is a CGo-free port of SQLite.  There's one file "picsync-metadata-cache.db" that contains
the cache, and if anything goes wrong, we'll just throw it away (since it is just a cache).

So far, the implementation can create the database and set up the schema.