Don't download again if it is in the cache
Problem: On 'picsync googlephoto list --update-cache' we were always downloading images and recalculating checksums even for things for which we already knew the checksum.

Solution: Add GetGooglephoto which lets us get an entry from the cache if it exists.  Use it in the update function - once we have the baseUrl, see if it is already in the cache (and update LRU time) before we download it all over again.