Remove useless googlephoto cobra commands
Problem: I defined a function for the command "picsync googlephotos"
even though all it does is print an error and exit (because you
need to choose a subcommand).  This is useless and non-standard usage
of cobra.  Also it says in the help for this set of commands that
this will sync from googlephotos to nixplay, but it won't.  That
functionality doesn't exist yet, and it'll probably live under "picsync
sync" to look like the smugmug->nixplay syncing.

Solution: Fix help text, and remove the command.  Now cobra prints
perfectly useful help if someone types "picsync googlephotos".