Add "picsync nixplay list" command
Problem: No list-only commands for nixplay are exposed (even though the
code is there).  This is useful for making sure nixplay stuff is
working, and will help with populating the nixplay cache more easily
(coming next).

Solution: Add a command, similar to the googlephotos list command.  It
currently lists albums but not photos for an album, yet.