Use google photos ID, not BaseUrl as cache key
Problem: The Google Photos BaseUrl changes for a particular photo.  It
is true that it never contains different content, but a particular
photo's BaseUrl will change (I think it may include a short-lived
authentication token).  So maintaining the google photos cache using the
BaseUrl means that the cache is never wrong, but not-very-useful, as
every few hours or so, it becomes totally invalidated.

Solution: Google photos publishes an Id that they guarantee never
changes:

    id:   A permanent, stable ID used to identify the object.

Use that as the cache key everywhere.  We'll still store baseUrl as a
column but it won't be useful as a key. It will probably frequently be
overwritten.