Delete old photos and use v3 (playlist) API
Problem: Nixplay has enhanced their API since this was originally written.  Originally, the "group of photos to show on a frame" was called a slideshow, and it is now called a playlist.  Old slideshows were automatically "converted" but with some weird behaviors (deletion doesn't appear to work right, and the CoverUrls are different).

(note neither a slideshow or a playlist is the same as an album.  I don't understand the UX purpose for the extra layer of indirection, but it is what it is.  We basically just try to get the Album right, then put 100% of an album into a similarly-named slideshow or playlist, ignoring the extra indirection)

Solution: Start using the new Playlist API by adding create, find, publish methods.  Then use them in the Googlephotos Sync path.

Probably we should deprecate and remove the smugmug capability soon.