Follow instructions from config file
Problem: The syntax for syncing was getting unwieldy, especially as soon
I may want to sync from multiple albums to multiple albums.

Solution: Put the sync instructions into a config file that is in YAML.  An example is in picsync.yaml.  Now the command is:
    picsync sync [picsync.yaml]
(you can specify a different config as the arg if you want)

I moved the previous picsync-config.yaml to .picsync-credentials.yaml and left it alone.  That's where passwords and API keys and stuff should live, that you should never share.  But the rest of the config is OK to share.