Serve cache stats via prometheus
Problem: No easy way to monitor long-term syncinc

Solution: Add prometheus integration.

This follows a depedency-injection pattern.  The packages get a prometheus
registerer passed as an argument and don't worry about how it is served.
This is a little more complicated than just using the global registry but it will
let us pass different registries in the future (especially in unit tests, where
you may want to pass mock registries to easily count things).

At the top level, promInitOrDie() actually just uses the default registry.

If the user doesn't pass a prometheus option in picsync.yaml, then we
simply don't init the prometheus stuff.  The registerer is nil, so all of the
metrics are still created but not hooked up to anything (metrics are not
heavyweight in prometheus).