Internal changes to allow prometheus in googlephotos client
Problem: No prometheus metrics from googlephotos client

Solution: Restructure this package and add prometheus infrastructure.
No metrics yet.

Previously, we were returning a *http.Client with OAuth2 hooks, since that
was all we needed.  We're spending a lot of effort juggling the http
client, the cache, and now the prometheus registry.  Just hide it all in
googlephotos.clientImpl and change the global-getters that took the
*http.Client into methods on the googlephotos.Client interface.