Report token expiry as "valid time remaining" instead
Problem: Reporting token expiry as a float64 is awkward to work with.
Plotters like grafana would have to convert it into a timestamp and diff
it from their local time which may produce confusing results and is easy
to get wrong.

Solution: Report how long we, the server, think is valid in remaining
token lifetime.  This will be easier to work with - as long as this is a
positive number, then everything is good.  In reality, probably expect
it to be more than 60 (as in, valid for more than one more minute).  If
it ever becomes negative, you should expect errors fetching google
photos.