Always get API keys from credential file, not env
Problem: We allowed setting some of the credentials (like the
less-sensitive google API key/secret) via env, and others via the
credentials file.  This is hard to support in Kubernetes (where we'd
have to bind some secrets to env vars, and others to files; result is
more secrets than we should ask the user to manager), and confusing
elsewhere.

Solution: Consolidate - all of these are set in
.picsync-credentials.yaml, either in the local directory or in
/etc/picsync-credentials as a fallback (we'll use that in kubernetes).