Download full-res image from Google Photos
Problem: Picsync was not appending anything to the baseUrl to indicate
what size of image to download.  In this case, google does not give you
full-res, it gives you some smaller res version (bigger than thumbnail,
but not great).  This was causing the images on the frame to look
pixelated.

Solution: Download full image by appending the "=d" (download) parameter
to the baseUrl.  As well, start tracking the image size (WxH) in the
cache (we could use this instead of "=d" if the EXIF data messes
anything up, like md5s, but we probably won't need it).