Close HTTP response bodies when downloading google images
Problem: High memory consumption when we have an empty cache or have to
upload google photos to nixplay.

Solution: In the cache case, we were not closing response body; we
should close it.

In the download->upload, we were closing but the defer statement was a
little later than it should have been (we would fail to
close if the response was an error HTTP response); this didn't happen
and so it isn't the cause of the leak, but good to improve anyway.

For uploads, the HTTP standard lib says it closes the input body (as
long as it is an io.Closer, which it is).