Expand Prometheus, including expected operating behavior
Problem: Prometheus monitoring section was small and outdated.

Solution: Dumping all the metrics in the top-level README is too
verbose.  Pull the metrics out, and put them in a separate doc that we
link to.  Also expand as we now have 40 total metrics.

Include a snapshot from grafana hooked up to prometheus that I got
earlier.

Add a section "Expected Behavior" that explains what a user should
expect to see when monitoring.