Clarify that we support syncing from shared albums
Problem: MaxWinterstein pointed out that this is one of the big
differences between picsync and the capability built into the nixplay
web app.  But it's not very clear from the docs.

Solution: Clarify docs to make this easy to identify.

- In the picsync.yaml example, clarify that it can be a shared album ID.
Also clarify that you can have multiple source album IDs for one nixplay
target.

- Document the `picsync googlephotos list --shared` command variant.

- In the "Comparison to Nixplay Built-In" section, describe this
difference.