Fix #5: Could not create playlist
Problem: When nixplay has to create a playlist, this error message is
always returned:

    Error syncing album Alle: couldn't create playlist ss_Alle: http 400: {"code":400,"message":"name (string) is required","name":"AssertionError"}

Solution: This is because we are not setting "content-type:
application/json" in the request, so the Nixplay API does not know which
decoder to apply to parse the request body.  Set "content-type:
application/json" and it works.

Meanwhile, also add code to create the album if needed because that's
another unfriendly step we require of first-time users.

Add prometheus counters for album creation.