Fix #4: googlephotos says "Must provide a Google Photos API key"
Problem: Whenever trying to do anything with Google Photos (googlephotos
login being the first step if following the setup instructions), you get
an error:

    panic: Must provide a Google Photos API key

Solution: I had avoided documenting this as I was hoping to apply for a
permanent key, but was also wanting to implement narrower permission
scopes (like read-only for a particular album rather than read-only all
albums), and didn't do that yet.

My mistake as this bug bit the first user to try it out.

For now I'm adding documentation about how to apply for a client
ID/secret, which is not as cool as having one already, but will be
temporarily OK.