Add --cache option
Problem: In Kubernetes, we'll need to keep the cache on a
PersistentVolume for it to be effective.  Currently picsync just creates
its cache file in the current working directory.  We could change
kubernetes working directory or do cd commands or something but this
makes it awkward if the user doesn't want to have a cache (we'd have to
have multiple example deployment files).  Probably easiest to just allow
an arg specifying where the cache file is.

Solution: Add "--cache" flag to the top-level command (so it is
available for all subcommands).