Do not treat "failure to list album" as "album does not exist"
Problem: If an error (like a network timeout, HTTP error or API
mismatch) is encountered when listing albums, the sync code will assume
that means the album does not exist and will attempt to create it.  This
may succeed.  If that succeeds, then there will be two albums created
with the same name that are essentially duplicates of each other.

In the case of repeated errors, this can repeat indefinitely, leading to
thousands of duplicate albums.

Solution: Do not treat "failure to list album" as "album does not
exist".  There are now four conditions to account for:

1. Failure to list albums (like network error): fail this operation
(which means we will retry after timeout if we are in the sync-forever
mode).

2. Successfully listed albums but could not find the one the caller
requested (length of matching albums is 0): we should try to create a
new album with that name, and proceed with sync.

3. Successfully listed albums and found exactly one match: proceed with
sync.

4. Successfully listed albums and found more than one match: fail this
operation and tell the user to delete the extra albums.