Fix #9: Implement mandatory pagination for Nixplay
Problem: Listing photos from nixplay now always fails with a 503 error.

Solution: It appears the nixplay API for listing photos has been updated
in two ways:

  1. The URI for listing photos is "/album/{id}/pictures/json/" not
"/album/{id}/pictures/json" (trailing slash is now required).

  2. You must implement paging, and pass "?page=1&limit=500" or
equivalent.

To accomodate this, we now use paging wherever we are listing albums.  I
tried it with small page sizes (10) to make sure the logic works, and
then bumped the limit up to 100 to avoid too-small HTTP transactions
(the web app uses 500).

The item limit is not currently configurable, because I'm not sure users
would need to set it.  For the sync to work, there's not really much
choice aside from "list all the pictures" (even if it takes multiple
transactions).