Update dependencies to avoid CVE-2022-32149
Found in Github Security Advisory:

https://github.com/advisories/GHSA-69ch-w2m2-3vjp

I don't think this functionality is used in picsync or probably any of
its dependencies.

The issue is that an attacker could cause a vulnerable program to spend
quadratic time parsing some headers.  We only communicate over HTTPS, so
this would mean an attacker had already compromised Google Photos or
Nixplay, and the worst case would be slow syncs.

Even though this is unlikely to matter, it's good to update and avoid
the vulnerability (and it's been a while since picsync deps were updated
at all).